Change the Context switching to be more flexible than the stack based switching.

+++ src/com/j2igf/framework/graphics/visual/Sprite.java
@@ -196,8 +196,8 @@ public class Sprite {
             Debug.logError(getClass().getName() + " -> Renderer instance can not be null!");
             System.exit(-1);
         }
-        x -= originX * width;
-        y -= originY * height;
+        x -= (int) (originX * width);
+        y -= (int) (originY * height);
         int startX = 0, startY = 0;
         int endX = width, endY = height;
